feat: modularizar ProductForm y completar aplicación de gestión de productos
